requirements.txt
appnope==0.1.2
azure-core==1.21.1
azure-storage-blob==12.9.0
backcall==0.2.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
idna==3.3
ipykernel==6.6.1
ipython==7.31.0
isodate==0.6.1
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
matplotlib-inline==0.1.3
msrest==0.6.21
mypy-extensions==0.4.3
mysql-connector==2.2.9
nest-asyncio==1.5.4
numpy==1.22.0
oauthlib==3.1.1
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.11.2
pyodbc==4.0.32
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
